clean up world and world activity classes
